Split-off deserialization code from Document
